Updated .NET version
