+++ TelCo.ColorCoder/Colors.cs
@@ -2,7 +2,6 @@
 
 namespace TelCo.ColorCoder
 {
-
     /// <summary>
     /// data type defined to hold the two colors of clor pair
     /// </summary>
@@ -15,7 +14,6 @@ namespace TelCo.ColorCoder
             return string.Format("MajorColor:{0}, MinorColor:{1}", majorColor.Name, minorColor.Name);
         }
     }
-
     internal static class ColorMaps
     {
         /// <summary>
@@ -26,8 +24,6 @@ namespace TelCo.ColorCoder
         /// Array of minor colors
         /// </summary>
         public static Color[] colorMapMinor { get; set; }
-
-  
         /// <summary>
         /// Static constructor required to initialize static variable
         /// </summary>
@@ -37,7 +33,6 @@ namespace TelCo.ColorCoder
             colorMapMajor = new Color[] { Color.White, Color.Red, Color.Black, Color.Yellow, Color.Violet };
             colorMapMinor = new Color[] { Color.Blue, Color.Orange, Color.Green, Color.Brown, Color.SlateGray };
         }
-
     }
 }
