+++ TelCo.ColorCoder/TelCo.ColorCoder.csproj
@@ -6,7 +6,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.CodeAnalysis.Metrics" Version="3.3.0" />
+    <PackageReference Include="Microsoft.CodeAnalysis.Metrics" Version="3.1.0" />
   </ItemGroup>
 
 </Project>
