+++ build.bat
@@ -1,5 +1,5 @@
 dotnet restore
-dotnet add TelCo.ColorCoder package Microsoft.CodeAnalysis.Metrics --version 2.1.0
+dotnet add TelCo.ColorCoder package Microsoft.CodeAnalysis.Metrics 
 pushd "%~dp0"
 if exist Debug rd /s /q Debug
 if exist Release rd /s /q Release
